[MIG] web_group_expand: Migration to 16.0
