+++ web_group_expand/README.rst
@@ -7,7 +7,7 @@ Group Expand Buttons
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:1e707b0c988c388ee8ae35b00354343f1e961eb524838d3fd56348df767e4ac8
+   !! source digest: sha256:0fb42d09f0995520749d0cb55db14e5ed47e9e037598f2a9ac9e5534887e0666
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -72,6 +72,7 @@ Contributors
 * Jan Verbeek <jverbeek@therp.nl>
 * Manuel Calero <manuelcalerosolis@gmail.com>
 * Alvaro Estebanez (brain-tec AG) <alvaro.estebanez@bt-group.com>
+* Mayank Patel <mayankpatel3555@gmail.com>
 
 Maintainers
 ~~~~~~~~~~~
