+++ web_group_expand/__manifest__.py
@@ -1,7 +1,7 @@
 {
     "name": "Group Expand Buttons",
     "category": "Web",
-    "version": "15.0.1.0.0",
+    "version": "16.0.1.0.0",
     "license": "AGPL-3",
     "author": "OpenERP SA, "
     "AvanzOSC, "
@@ -13,10 +13,8 @@
     "depends": ["web"],
     "assets": {
         "web.assets_backend": [
-            "/web_group_expand/static/src/js/web_group_expand.esm.js",
-        ],
-        "web.assets_qweb": [
-            "/web_group_expand/static/src/xml/expand_buttons.xml",
+            "/web_group_expand/static/src/xml/list_controller.xml",
+            "/web_group_expand/static/src/js/list_controller.esm.js",
         ],
     },
 }
