+++ web_group_expand/i18n/ca.po
@@ -17,15 +17,23 @@ msgstr ""
 "X-Generator: Weblate 4.3.2\n"
 
 #. module: web_group_expand
-#. openerp-web
-#: code:addons/web_group_expand/static/src/xml/expand_buttons.xml:0
+#. odoo-javascript
+#: code:addons/web_group_expand/static/src/xml/list_controller.xml:0
 #, python-format
-msgid "Collapse groups"
-msgstr "Col·lapsar grups"
+msgid "Compress"
+msgstr ""
 
 #. module: web_group_expand
-#. openerp-web
-#: code:addons/web_group_expand/static/src/xml/expand_buttons.xml:0
+#. odoo-javascript
+#: code:addons/web_group_expand/static/src/xml/list_controller.xml:0
 #, python-format
-msgid "Expand groups"
-msgstr "Expandir grups"
+msgid "Expand"
+msgstr ""
+
+#, python-format
+#~ msgid "Collapse groups"
+#~ msgstr "Col·lapsar grups"
+
+#, python-format
+#~ msgid "Expand groups"
+#~ msgstr "Expandir grups"
