+++ web_group_expand/i18n/web_group_expand.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 15.0\n"
+"Project-Id-Version: Odoo Server 16.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -14,15 +14,15 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. module: web_group_expand
-#. openerp-web
-#: code:addons/web_group_expand/static/src/xml/expand_buttons.xml:0
+#. odoo-javascript
+#: code:addons/web_group_expand/static/src/xml/list_controller.xml:0
 #, python-format
-msgid "Collapse groups"
+msgid "Compress"
 msgstr ""
 
 #. module: web_group_expand
-#. openerp-web
-#: code:addons/web_group_expand/static/src/xml/expand_buttons.xml:0
+#. odoo-javascript
+#: code:addons/web_group_expand/static/src/xml/list_controller.xml:0
 #, python-format
-msgid "Expand groups"
+msgid "Expand"
 msgstr ""
