+++ web_group_expand/readme/CONTRIBUTORS.rst
@@ -5,3 +5,4 @@
 * Jan Verbeek <jverbeek@therp.nl>
 * Manuel Calero <manuelcalerosolis@gmail.com>
 * Alvaro Estebanez (brain-tec AG) <alvaro.estebanez@bt-group.com>
+* Mayank Patel <mayankpatel3555@gmail.com>
