+++ web_group_expand/static/description/index.html
@@ -367,7 +367,7 @@ ul.auto-toc {
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-!! source digest: sha256:1e707b0c988c388ee8ae35b00354343f1e961eb524838d3fd56348df767e4ac8
+!! source digest: sha256:0fb42d09f0995520749d0cb55db14e5ed47e9e037598f2a9ac9e5534887e0666
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
 <p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/OCA/web/tree/16.0/web_group_expand"><img alt="OCA/web" src="https://img.shields.io/badge/github-OCA%2Fweb-lightgray.png?logo=github" /></a> <a class="reference external image-reference" href="https://translation.odoo-community.org/projects/web-16-0/web-16-0-web_group_expand"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external image-reference" href="https://runboat.odoo-community.org/builds?repo=OCA/web&amp;target_branch=16.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
 <p>When grouping a list by a field, this module adds two buttons to expand or
@@ -416,6 +416,7 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <li>Jan Verbeek &lt;<a class="reference external" href="mailto:jverbeek&#64;therp.nl">jverbeek&#64;therp.nl</a>&gt;</li>
 <li>Manuel Calero &lt;<a class="reference external" href="mailto:manuelcalerosolis&#64;gmail.com">manuelcalerosolis&#64;gmail.com</a>&gt;</li>
 <li>Alvaro Estebanez (brain-tec AG) &lt;<a class="reference external" href="mailto:alvaro.estebanez&#64;bt-group.com">alvaro.estebanez&#64;bt-group.com</a>&gt;</li>
+<li>Mayank Patel &lt;<a class="reference external" href="mailto:mayankpatel3555&#64;gmail.com">mayankpatel3555&#64;gmail.com</a>&gt;</li>
 </ul>
 </div>
 <div class="section" id="maintainers">
