--- web_group_expand/static/src/xml/expand_buttons.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates>
-    <t t-name="web_group_expand.Buttons">
-        <nav class="btn-group oe_group_by_expand_buttons o_hidden">
-            <button
-                class="btn btn-secondary fa fa-expand"
-                id="oe_group_by_expand"
-                title="Expand groups"
-                type="button"
-            />
-            <button
-                class="btn btn-secondary fa fa-compress"
-                id="oe_group_by_collapse"
-                title="Collapse groups"
-                type="button"
-            />
-        </nav>
-    </t>
-</templates>
